build route POST /meals
hard codes calorie_goal until the goal table can be querried. It is understood that the reason why Travis is failing is because in their test db, they are trying to add items with an id of 1, which is conficting with seeds. It is passing locally because that id is never getting reset